From CDI to EJB
