+++ pom.xml
@@ -95,6 +95,15 @@
       <scope>runtime</scope>
     </dependency>
 
+    <!-- https://mvnrepository.com/artifact/org.jboss.spec.javax.ejb/jboss-ejb-api_3.2_spec -->
+    <dependency>
+      <groupId>org.jboss.spec.javax.ejb</groupId>
+      <artifactId>jboss-ejb-api_3.2_spec</artifactId>
+      <version>2.0.0.Final</version>
+      <scope>provided</scope>
+    </dependency>
+
+
   </dependencies>
 
   <build>
