+++ src/main/java/ru/geekbrains/shop/controller/ProductController.java
@@ -3,8 +3,8 @@ package ru.geekbrains.shop.controller;
 import ru.geekbrains.shop.persist.Product;
 import ru.geekbrains.shop.persist.repository.ProductRepository;
 
+import javax.ejb.EJB;
 import javax.enterprise.context.SessionScoped;
-import javax.inject.Inject;
 import javax.inject.Named;
 import java.io.Serializable;
 import java.util.List;
@@ -13,7 +13,7 @@ import java.util.List;
 @SessionScoped
 public class ProductController implements Serializable {
 
-    @Inject
+    @EJB
     private ProductRepository productRepository;
 
     private Product product;
